+++ Changelog.md
@@ -7,7 +7,7 @@
 - `saveXYZfromTrajGroup` and `getXYZfromTrajGroup` now can export comments per frames
 - `transitionMatrixFromSOAPClassification` now creates matrix with shape  `(n,n)` and no more `(n+1,n+1)`, where `n` is the lenght of the legend. The user will now need to address the errors in classification, if needed
 - added `calculateResidenceTimesFromClassification` for calculating the residence times of the states during the MD
-- added `calculateEvents` for calculating the history of the evolution of the states in the MD
+- added `trackStates` for calculating the history of the evolution of the states in the MD
 
 ## Changes From 0.0.2a